introduce ZRB_INIT_MODULES
