+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "zrb"
-version = "0.24.1"
+version = "0.25.0"
 description = "A Framework to Enhance Your Workflow"
 authors = ["Go Frendi Gunawan <gofrendiasgard@gmail.com>"]
 license = "AGPL-3.0-or-later"
