+++ src/zrb/helper/cli.py
@@ -4,9 +4,10 @@ from functools import lru_cache
 
 import click
 
-from zrb.config.config import INIT_SCRIPTS, SHOULD_LOAD_BUILTIN, VERSION
+from zrb.config.config import INIT_MODULES, INIT_SCRIPTS, SHOULD_LOAD_BUILTIN, VERSION
 from zrb.helper.accessories.color import colored
 from zrb.helper.loader.load_module import load_module
+from zrb.helper.loader.load_script import load_script
 from zrb.helper.log import logger
 from zrb.helper.typecheck import typechecked
 from zrb.runner import runner
@@ -45,15 +46,28 @@ def create_cli() -> click.Group:
         from zrb import builtin
 
         assert builtin
+    # load from ZRB_INIT_MODULES
+    for init_module in INIT_MODULES:
+        try:
+            load_module(init_module)
+        except Exception:
+            logger.error(
+                colored(
+                    f"Failed to load module {init_module}",
+                    color="red",
+                    attrs=["bold"],
+                )
+            )
+            traceback.print_exc()
     # Load zrb_init.py
     project_dir = os.getenv("ZRB_PROJECT_DIR", os.getcwd())
     project_script = os.path.join(project_dir, "zrb_init.py")
-    load_module(script_path=project_script, sys_path_index=0)
+    load_script(script_path=project_script, sys_path_index=0)
     # load from ZRB_INIT_SCRIPTS environment
     for index, init_script in enumerate(INIT_SCRIPTS):
         logger.info(colored(f"Load module from {init_script}", attrs=["dark"]))
         try:
-            load_module(script_path=init_script, sys_path_index=index + 1)
+            load_script(script_path=init_script, sys_path_index=index + 1)
         except Exception:
             logger.error(
                 colored(
